# Conflicts:
#	core/src/main/java/io/seata/core/rpc/RpcContext.java
#	core/src/main/java/io/seata/core/rpc/netty/NettyClientChannelManager.java
#	core/src/main/java/io/seata/core/rpc/netty/NettyServerBootstrap.java
#	core/src/main/java/io/seata/core/rpc/netty/RpcServer.java

+++ common/src/main/java/io/seata/common/util/BlobUtils.java
@@ -25,7 +25,7 @@ import io.seata.common.exception.ShouldNeverHappenException;
 /**
  * The type Blob utils.
  *
- * @author jimin.jm @alibaba-inc.com
+ * @author slievrly
  * @author Geng Zhang
  */
 public class BlobUtils {
@@ -64,7 +64,7 @@ public class BlobUtils {
         }
 
         try {
-            return new String(blob.getBytes((long) 1, (int) blob.length()), Constants.DEFAULT_CHARSET);
+            return new String(blob.getBytes(1, (int) blob.length()), Constants.DEFAULT_CHARSET);
         } catch (Exception e) {
             throw new ShouldNeverHappenException(e);
         }
@@ -100,7 +100,7 @@ public class BlobUtils {
         }
 
         try {
-            return blob.getBytes((long) 1, (int) blob.length());
+            return blob.getBytes(1, (int) blob.length());
         } catch (Exception e) {
             throw new ShouldNeverHappenException(e);
         }
